ports_set

+++ srcs/verbose.c
@@ -2,7 +2,7 @@
 
 void print_hostnames(t_nmap* nmap) {
     printf("Hosts: ");
-    for (size_t i = 0; i < nmap->hostname_count; ++i) printf(i == 0 ? "%s" : ", %s", nmap->hostnames[i]);
+    for (int i = 0; i < nmap->hostname_count; ++i) printf(i == 0 ? "%s" : ", %s", nmap->hostnames[i]);
     printf("\n");
 }
